administrace zařízení

+++ backend/api/auth/delete.js
@@ -1,5 +1,7 @@
 const router = require('express').Router()
 const auth = require('../../modules/auth')
+let { DB } = require('mongquick')
+const db = new DB(process.env.MongoLogin)
 
 router.post('/', async (req, res) => {
 	try {
@@ -14,7 +16,7 @@ router.post('/', async (req, res) => {
 		return res
 			.status(409)
 			.send({ valid: false, response: 'Uživatel s tímto jménem nexistuje' })
-
+if (!auth.authenticatePassword(req.body.password, await db.get(req.body.username).user.password)) return res.status(401).send({ valid: false, response: 'heslo není správné' })
 	auth.deleteUser(req.body.username)
 	res.status(200).send({ valid: true })
 }
